# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Coupis API",
    "description": "API for biodiversity occurrence exploration.",
    "version": "0.1.0"
  },
  "paths": {
    "/health": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/occurrences/temporal-extent": {
      "get": {
        "tags": [
          "occurrences"
        ],
        "summary": "Get Occurrence Temporal Extent",
        "operationId": "get_occurrence_temporal_extent_api_v1_occurrences_temporal_extent_get",
        "parameters": [
          {
            "name": "species_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Species Id"
            }
          },
          {
            "name": "region_slug",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Region Slug"
            }
          },
          {
            "name": "region_version",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Region Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OccurrenceTemporalExtentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/occurrences/yearly-counts": {
      "get": {
        "tags": [
          "occurrences"
        ],
        "summary": "Get Occurrence Yearly Counts",
        "operationId": "get_occurrence_yearly_counts_api_v1_occurrences_yearly_counts_get",
        "parameters": [
          {
            "name": "species_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Species Id"
            }
          },
          {
            "name": "region_slug",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Region Slug"
            }
          },
          {
            "name": "region_version",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Region Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OccurrenceYearlyCountsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/occurrences": {
      "get": {
        "tags": [
          "occurrences"
        ],
        "summary": "List Occurrences",
        "operationId": "list_occurrences_api_v1_occurrences_get",
        "parameters": [
          {
            "name": "species_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Species Id"
            }
          },
          {
            "name": "gbif_taxon_key",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Gbif Taxon Key"
            }
          },
          {
            "name": "dataset_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Dataset Id"
            }
          },
          {
            "name": "region_slug",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Region Slug"
            }
          },
          {
            "name": "region_version",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Region Version"
            }
          },
          {
            "name": "bbox",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Bbox"
            }
          },
          {
            "name": "observed_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Observed From"
            }
          },
          {
            "name": "observed_until",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Observed Until"
            }
          },
          {
            "name": "basis_of_record",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Basis Of Record"
            }
          },
          {
            "name": "occurrence_status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Occurrence Status"
            }
          },
          {
            "name": "country_code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 2,
                  "maxLength": 2
                },
                {
                  "type": "null"
                }
              ],
              "title": "Country Code"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OccurrencePageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/regions": {
      "get": {
        "tags": [
          "regions"
        ],
        "summary": "List Regions",
        "operationId": "list_regions_api_v1_regions_get",
        "parameters": [
          {
            "name": "include_all_versions",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include All Versions"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RegionSummaryResponse"
                  },
                  "title": "Response List Regions Api V1 Regions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/regions/{slug}": {
      "get": {
        "tags": [
          "regions"
        ],
        "summary": "Get Region",
        "operationId": "get_region_api_v1_regions__slug__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          },
          {
            "name": "version",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegionDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/species": {
      "get": {
        "tags": [
          "species"
        ],
        "summary": "List Species",
        "operationId": "list_species_api_v1_species_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 200
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "gbif_taxon_key",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Gbif Taxon Key"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpeciesPageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "GeoJSONMultiPolygonGeometry": {
        "properties": {
          "type": {
            "type": "string",
            "const": "MultiPolygon",
            "title": "Type"
          },
          "coordinates": {
            "items": {
              "items": {
                "items": {
                  "items": {
                    "type": "number"
                  },
                  "type": "array"
                },
                "type": "array"
              },
              "type": "array"
            },
            "type": "array",
            "title": "Coordinates"
          }
        },
        "type": "object",
        "required": [
          "type",
          "coordinates"
        ],
        "title": "GeoJSONMultiPolygonGeometry"
      },
      "GeoJSONPolygonGeometry": {
        "properties": {
          "type": {
            "type": "string",
            "const": "Polygon",
            "title": "Type"
          },
          "coordinates": {
            "items": {
              "items": {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              "type": "array"
            },
            "type": "array",
            "title": "Coordinates"
          }
        },
        "type": "object",
        "required": [
          "type",
          "coordinates"
        ],
        "title": "GeoJSONPolygonGeometry"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HealthResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "HealthResponse"
      },
      "OccurrencePageResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/OccurrenceResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "OccurrencePageResponse"
      },
      "OccurrenceResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "gbif_id": {
            "type": "integer",
            "title": "Gbif Id"
          },
          "occurrence_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occurrence Id"
          },
          "species_id": {
            "type": "integer",
            "title": "Species Id"
          },
          "dataset_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset Id"
          },
          "observed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observed At"
          },
          "latitude": {
            "type": "number",
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "title": "Longitude"
          },
          "coordinate_uncertainty_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coordinate Uncertainty M"
          },
          "basis_of_record": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Basis Of Record"
          },
          "occurrence_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occurrence Status"
          },
          "country_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country Code"
          },
          "locality": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Locality"
          },
          "license": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "issues": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issues"
          }
        },
        "type": "object",
        "required": [
          "id",
          "gbif_id",
          "occurrence_id",
          "species_id",
          "dataset_id",
          "observed_at",
          "latitude",
          "longitude",
          "coordinate_uncertainty_m",
          "basis_of_record",
          "occurrence_status",
          "country_code",
          "locality",
          "license",
          "source_url",
          "issues"
        ],
        "title": "OccurrenceResponse"
      },
      "OccurrenceTemporalExtentResponse": {
        "properties": {
          "observed_from": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observed From"
          },
          "observed_until": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observed Until"
          }
        },
        "type": "object",
        "required": [
          "observed_from",
          "observed_until"
        ],
        "title": "OccurrenceTemporalExtentResponse"
      },
      "OccurrenceYearCountResponse": {
        "properties": {
          "year": {
            "type": "integer",
            "title": "Year"
          },
          "count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "year",
          "count"
        ],
        "title": "OccurrenceYearCountResponse"
      },
      "OccurrenceYearlyCountsResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/OccurrenceYearCountResponse"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "OccurrenceYearlyCountsResponse"
      },
      "RegionDetailResponse": {
        "properties": {
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "license": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License"
          },
          "geometry": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/GeoJSONPolygonGeometry"
              },
              {
                "$ref": "#/components/schemas/GeoJSONMultiPolygonGeometry"
              }
            ],
            "title": "Geometry",
            "discriminator": {
              "propertyName": "type",
              "mapping": {
                "MultiPolygon": "#/components/schemas/GeoJSONMultiPolygonGeometry",
                "Polygon": "#/components/schemas/GeoJSONPolygonGeometry"
              }
            }
          }
        },
        "type": "object",
        "required": [
          "slug",
          "name",
          "version",
          "source",
          "license",
          "geometry"
        ],
        "title": "RegionDetailResponse"
      },
      "RegionSummaryResponse": {
        "properties": {
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "license": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License"
          },
          "geometry": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/GeoJSONPolygonGeometry"
              },
              {
                "$ref": "#/components/schemas/GeoJSONMultiPolygonGeometry"
              }
            ],
            "title": "Geometry",
            "discriminator": {
              "propertyName": "type",
              "mapping": {
                "MultiPolygon": "#/components/schemas/GeoJSONMultiPolygonGeometry",
                "Polygon": "#/components/schemas/GeoJSONPolygonGeometry"
              }
            }
          }
        },
        "type": "object",
        "required": [
          "slug",
          "name",
          "version",
          "source",
          "license",
          "geometry"
        ],
        "title": "RegionSummaryResponse"
      },
      "SpeciesPageResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SpeciesResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "SpeciesPageResponse"
      },
      "SpeciesResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "gbif_taxon_key": {
            "type": "integer",
            "title": "Gbif Taxon Key"
          },
          "scientific_name": {
            "type": "string",
            "title": "Scientific Name"
          },
          "canonical_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canonical Name"
          },
          "vernacular_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vernacular Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "gbif_taxon_key",
          "scientific_name",
          "canonical_name",
          "vernacular_name"
        ],
        "title": "SpeciesResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}